# openapi.yml
openapi: 3.0.3
info:
  title: Space Marine Management API
  version: 1.0.0
  description: API for managing Space Marines, their Coordinates, and Chapters with full CRUD, pagination support, special operations, embedded entities, and export capabilities.
servers:
  - url: http://localhost:8080/api
    description: Local development server
paths:
...
  /chapters:
    get:
      summary: Get all Chapters (paginated)
      operationId: getAllChapters
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          description: Zero-based page index
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page (max 100)
      responses:
        "200":
          description: Paginated list of Chapters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedChapterResponse"
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create a new Chapter
      operationId: createChapter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Chapter"
      responses:
        "201":
          description: Chapter created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterWithId"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chapters/{id}:
    get:
      summary: Get a Chapter by ID
      operationId: getChapterById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Chapter found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterWithId"
        "404":
          description: Chapter not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Update a Chapter by ID
      operationId: updateChapter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Chapter"
      responses:
        "200":
          description: Chapter updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterWithId"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Chapter not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete a Chapter by ID
      operationId: deleteChapter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "204":
          description: Chapter deleted
        "404":
          description: Chapter not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chapters/export/json:
    get:
      summary: Export Chapters as JSON
      operationId: exportChaptersJson
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          description: Zero-based page index
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 1000
            default: 1000
          description: Number of items per page (max 1000 for exports)
      responses:
        "200":
          description: JSON file download with paginated Chapters data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChapterWithId"
          headers:
            Content-Disposition:
              description: Attachment filename
              schema:
                type: string
                example: "chapters_page_0_size_1000.json"
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chapters/export/xml:
    get:
      summary: Export Chapters as XML
      operationId: exportChaptersXml
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          description: Zero-based page index
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 1000
            default: 1000
          description: Number of items per page (max 1000 for exports)
      responses:
        "200":
          description: XML file download with paginated Chapters data
          content:
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChapterWithId"
          headers:
            Content-Disposition:
              description: Attachment filename
              schema:
                type: string
                example: "chapters_page_0_size_1000.xml"
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    PaginatedChapterResponse:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ChapterWithId"
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          format: int32
          minimum: 0

    # === EMBEDDED PAGINATION WRAPPER ===
...
    Chapter:
      type: object
      required:
        - name
        - marinesCount
      properties:
        name:
          type: string
          minLength: 1
          description: Name of the Chapter
        marinesCount:
          type: integer
          format: int64
          minimum: 1
          maximum: 1000
          description: Number of marines in the chapter (must be between 1 and 1000)
    ChapterWithId:
      allOf:
        - $ref: "#/components/schemas/Chapter"
        - type: object
          required:
            - id
          properties:
            id:
              type: integer
              format: int64
              minimum: 1
              description: Unique auto-generated ID

    # === ERROR RESPONSES ===
    ErrorResponse:
      type: object
      required:
        - error
        - timestamp
      properties:
        error:
          type: string
          description: Human-readable error message
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred
      example:
        error: "Invalid weapon types: LASER_CANNON. Valid values: BOLTGUN,HEAVY_BOLTGUN,FLAMER,HEAVY_FLAMER,MULTI_MELTA"
        timestamp: "2025-11-07T14:30:00.000Z"